Relocate workflow tasks to specific pkgs
