+++ pkg/workflow/task/bootstrap/bootstrap.go
@@ -1,4 +1,4 @@
-package task
+package bootstrap
 
 import (
 	"context"
